new commit

+++ Lesson 4/mini_project/profile.php
@@ -6,6 +6,7 @@ if (session_status() === PHP_SESSION_NONE) {
 <!DOCTYPE html>
 <html lang="en">
 <head>
+    <!-- ... your code ... -->
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
@@ -13,6 +14,7 @@ if (session_status() === PHP_SESSION_NONE) {
     <!-- Bootstrap CSS -->
     <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.0.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-EVSTQN3/azprG1Anm3QDgpJLIm9Nao0Yz1ztcQTwFspd3yD65VohhpuuCOmLASjC" crossorigin="anonymous">
     <link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.10.5/font/bootstrap-icons.css">
+    
     <style>
         table{
             width: 100%;
@@ -33,21 +35,31 @@ if (session_status() === PHP_SESSION_NONE) {
                 <th colspan='2'>Action</th>
             </tr>";
             include_once "database-config.php";
-            $user_logged_in = $_SESSION['username'];
-            $sql = "SELECT * FROM users WHERE username='$user_logged_in'";
-            $result = $database_connection->query($sql);
-            // var_dump($result);
-            $row = $result->fetch_assoc();
-            echo "<tr>
-                    <td>" . $row['username']. "</td>
-                    <td>".$row['email']."</td>
-                    <td>".$row['password']."</td>
-                    <td><a href='edit.php?id=$row[id]'><i class='bi bi-pencil-square text-success'></i></a></td>
-                    <td><a href='delete.php?id=$row[id]'><i class='bi bi-archive-fill text-danger'></i></a></td>
-                </tr>";
+            
+            if (isset($_SESSION['username'])) {
+                $user_logged_in = $_SESSION['username'];
+                $sql = "SELECT * FROM userProfile WHERE username='$user_logged_in'";
+                $result = $database_connection->query($sql);
+                if ($result && $result->num_rows > 0) {
+                    $row = $result->fetch_assoc();
+                    echo "<tr>
+                            <td>" . $row['username']. "</td>
+                            <td>".$row['email']."</td>
+                            <td>".$row['password']."</td>
+                            <td><a href='edit.php?id=$row[id]'><i class='bi bi-pencil-square text-success'></i></a></td>
+                            <td><a href='delete.php?id=$row[id]'><i class='bi bi-archive-fill text-danger'></i></a></td>
+                        </tr>";
+                } else {
+                    echo "<tr><td colspan='5'>No data available</td></tr>";
+                }
+            } else {
+                echo "<tr><td colspan='5'>User not logged in</td></tr>";
+            }
+            
         echo "</table>";
-?>
+    ?>
     </div>
+    <!-- ... your code ... -->
     <!-- Optional JavaScript; choose one of the two! -->
 
     <!-- Option 1: Bootstrap Bundle with Popper -->
